fix the bugs in model and database; add form to all teams tag in judging page
